+++ aws/static-website/CHANGELOG.md
@@ -6,8 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Features
 
-* Add `enable_logging` to enable logging from CloudFront to CloudWatch.
-* Add `log_bucket_arn` and `log_bucket_prefix_path` to manage log bucket destination for storing CloudFront logs.
+* Add `cloudfront_distribution_arn` to module `static-website` output.
+* Add module `cloudfront-logging` to deliver CloudFront logs to S3 log bucket.
 
 ## [0.6.4]() (2025-06-03)